requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
beautifulsoup4==4.8.2
black==19.10b0
bottle==0.12.18
cachelib==0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cssselect==1.1.0
Django==3.0.3
django-cors-headers==3.2.1
django-credit-cards==0.4.1
django-filter==2.2.0
django-safedelete==0.5.4
djangorestframework==3.11.0
emoji==0.5.4
feedfinder2==0.0.4
feedparser==5.2.1
geographiclib==1.50
geopy==1.20.0
howdoi==1.2.1
idna==2.8
imgcat==0.4.0
isort==4.3.21
jieba3k==0.35.1
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.18.1
pathspec==0.7.0
pep8==1.7.1
Pillow==7.0.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.22.0
requests-file==1.4.3
sh==1.12.14
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.7
virtualenv==16.7.9
wikipedia==1.4.0
wrapt==1.11.2
